requirements.txt
langchain==1.2.12
rich==14.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0